refactor(tauri): [#10] 백엔드 모듈 경계를 분리한다
