+++ src-tauri/src/application/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod archived_sessions;
+pub(crate) mod recent_sessions;
+pub(crate) mod workspace_identity;
