+++ src-tauri/src/application/recent_sessions.rs
@@ -1,0 +1,740 @@
+use crate::{
+    application::workspace_identity::resolve_live_session_workspace_identity,
+    domain::{
+        ingest_policy::{
+            is_supported_live_session_source, should_hide_recent_boot_thread, LIVE_SESSION_SOURCES,
+            MAX_RECENT_SESSIONS, RECENT_INDEX_PREFIX_SCAN_LIMIT, RECENT_INDEX_TAIL_BYTES,
+            RECENT_INDEX_TAIL_ENTRY_LIMIT,
+        },
+        session::{RecentSessionIndexItem, SessionLogSnapshot},
+        workspace::WorkspaceIdentity,
+    },
+    infrastructure::{
+        filesystem::{collect_jsonl_files, resolve_codex_home, resolve_projects_root},
+        session_jsonl::{
+            parse_live_session_snapshot, parse_recent_index_entry, read_subagent_snapshot,
+        },
+        state_sqlite::{load_live_thread_rows, LiveThreadRow},
+    },
+};
+use std::{
+    fs, io,
+    path::{Path, PathBuf},
+};
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+struct LiveSessionCandidate {
+    session_id: String,
+    file_path: PathBuf,
+    workspace_path: String,
+    workspace_identity: WorkspaceIdentity,
+}
+
+pub(crate) fn load_recent_session_index_from_disk() -> io::Result<Vec<RecentSessionIndexItem>> {
+    let codex_home = resolve_codex_home()?;
+    let projects_root = resolve_projects_root()?;
+    let sessions_root = codex_home.join("sessions");
+    let candidates = load_live_session_candidates(&codex_home, &sessions_root, &projects_root)?;
+
+    let mut items = Vec::new();
+
+    for candidate in candidates {
+        if items.len() >= MAX_RECENT_SESSIONS {
+            break;
+        }
+
+        let item = match build_recent_index_item(&candidate) {
+            Ok(Some(item)) => item,
+            Ok(None) => continue,
+            Err(_) => continue,
+        };
+        if should_hide_recent_boot_thread(&item) {
+            continue;
+        }
+
+        items.push(item);
+    }
+
+    Ok(items)
+}
+
+pub(crate) fn load_recent_session_snapshot_from_disk(
+    file_path: &str,
+) -> Option<SessionLogSnapshot> {
+    let codex_home = resolve_codex_home().ok()?;
+    let projects_root = resolve_projects_root().ok()?;
+    let sessions_root = codex_home.join("sessions");
+    let path = Path::new(file_path);
+
+    let canonical_path = fs::canonicalize(path).ok()?;
+    let canonical_root = fs::canonicalize(&sessions_root).ok()?;
+    if !canonical_path.starts_with(&canonical_root) {
+        return None;
+    }
+
+    let candidate = load_live_session_candidates(&codex_home, &sessions_root, &projects_root)
+        .ok()?
+        .into_iter()
+        .find(|item| item.file_path == canonical_path)?;
+    let mut snapshot =
+        build_recent_session_snapshot(&candidate.file_path, &projects_root).ok()??;
+
+    let mut session_files = Vec::new();
+    collect_jsonl_files(&sessions_root, &mut session_files).ok()?;
+
+    for session_file in &session_files {
+        if fs::canonicalize(session_file).ok().as_ref() == Some(&candidate.file_path) {
+            continue;
+        }
+
+        if let Ok(Some(subagent)) = read_subagent_snapshot(session_file) {
+            if subagent.parent_thread_id == snapshot.session_id
+                || snapshot
+                    .forked_from_id
+                    .as_ref()
+                    .is_some_and(|forked_from_id| subagent.parent_thread_id == *forked_from_id)
+            {
+                snapshot.subagents.push(subagent);
+            }
+        }
+    }
+
+    Some(snapshot)
+}
+
+fn load_live_session_candidates(
+    codex_home: &Path,
+    sessions_root: &Path,
+    projects_root: &Path,
+) -> io::Result<Vec<LiveSessionCandidate>> {
+    let rows = load_live_thread_rows(codex_home)?;
+    let canonical_sessions_root = fs::canonicalize(sessions_root)?;
+    let mut candidates = Vec::new();
+    for row in rows {
+        let Some(candidate) =
+            build_live_session_candidate(row, &canonical_sessions_root, projects_root)?
+        else {
+            continue;
+        };
+        candidates.push(candidate);
+    }
+
+    Ok(candidates)
+}
+
+fn build_live_session_candidate(
+    row: LiveThreadRow,
+    canonical_sessions_root: &Path,
+    projects_root: &Path,
+) -> io::Result<Option<LiveSessionCandidate>> {
+    if !is_supported_live_session_source(&row.source) {
+        return Ok(None);
+    }
+    if row.workspace_path.trim().is_empty() || row.rollout_path.trim().is_empty() {
+        return Ok(None);
+    }
+
+    let canonical_file_path = match fs::canonicalize(Path::new(&row.rollout_path)) {
+        Ok(path) => path,
+        Err(_) => return Ok(None),
+    };
+    if !canonical_file_path.starts_with(canonical_sessions_root) {
+        return Ok(None);
+    }
+
+    let workspace_identity = match resolve_live_session_workspace_identity(
+        Path::new(&row.workspace_path),
+        projects_root,
+    ) {
+        Ok(identity) => identity,
+        Err(_) => return Ok(None),
+    };
+
+    Ok(Some(LiveSessionCandidate {
+        session_id: row.session_id,
+        file_path: canonical_file_path,
+        workspace_path: row.workspace_path,
+        workspace_identity,
+    }))
+}
+
+fn build_recent_index_item(
+    candidate: &LiveSessionCandidate,
+) -> io::Result<Option<RecentSessionIndexItem>> {
+    let parsed = parse_recent_index_entry(
+        &candidate.file_path,
+        RECENT_INDEX_PREFIX_SCAN_LIMIT,
+        RECENT_INDEX_TAIL_BYTES,
+        RECENT_INDEX_TAIL_ENTRY_LIMIT,
+    )?;
+
+    Ok(parsed.map(|parsed| RecentSessionIndexItem {
+        session_id: candidate.session_id.clone(),
+        workspace_path: candidate.workspace_path.clone(),
+        origin_path: candidate.workspace_identity.origin_path.clone(),
+        display_name: candidate.workspace_identity.display_name.clone(),
+        started_at: parsed.started_at,
+        updated_at: parsed.updated_at,
+        model: parsed.model,
+        file_path: candidate.file_path.display().to_string(),
+        first_user_message: parsed.first_user_message,
+        title: parsed.title,
+        status: parsed.status,
+        last_event_summary: parsed.last_event_summary,
+    }))
+}
+
+fn build_recent_session_snapshot(
+    session_file: &Path,
+    projects_root: &Path,
+) -> io::Result<Option<SessionLogSnapshot>> {
+    let parsed = parse_live_session_snapshot(session_file, LIVE_SESSION_SOURCES)?;
+    let Some(parsed) = parsed else {
+        return Ok(None);
+    };
+
+    let workspace_identity =
+        resolve_live_session_workspace_identity(Path::new(&parsed.workspace_path), projects_root)
+            .ok();
+    let Some(workspace_identity) = workspace_identity else {
+        return Ok(None);
+    };
+
+    Ok(Some(SessionLogSnapshot {
+        session_id: parsed.session_id,
+        forked_from_id: parsed.forked_from_id,
+        workspace_path: parsed.workspace_path,
+        origin_path: workspace_identity.origin_path,
+        display_name: workspace_identity.display_name,
+        started_at: parsed.started_at,
+        updated_at: parsed.updated_at,
+        model: parsed.model,
+        entries: parsed.entries,
+        subagents: Vec::new(),
+        is_archived: false,
+        prompt_assembly: parsed.prompt_assembly,
+    }))
+}
+
+#[cfg(test)]
+mod tests {
+    use super::{load_recent_session_index_from_disk, load_recent_session_snapshot_from_disk};
+    use crate::{
+        domain::ingest_policy::DEFAULT_THREAD_TITLE,
+        test_support::{
+            create_git_workspace, create_linked_worktree, create_state_database, insert_thread_row,
+            insert_thread_row_with_archive_flag, session_meta_line, session_meta_line_with_fork,
+            session_meta_line_with_source, write_session_lines, RecentSessionTestContext,
+        },
+    };
+    use std::fs;
+
+    #[test]
+    fn builds_recent_index_entry_from_lightweight_session_scan() {
+        let ctx = RecentSessionTestContext::new("recent-index");
+        let workspace_path = ctx.projects_root.join("demo-app");
+        let session_file = ctx.sessions_root.join("session.jsonl");
+        let state_database = ctx.codex_home.join("state_1.sqlite");
+
+        create_git_workspace(&workspace_path);
+        create_state_database(&state_database, &[]);
+        write_session_lines(
+            &session_file,
+            vec![
+                session_meta_line("session-001", &workspace_path),
+                r#"{"timestamp":"2026-03-20T00:00:01.000Z","type":"turn_context","payload":{"model":"gpt-5"}}"#
+                    .to_owned(),
+                r#"{"timestamp":"2026-03-20T00:00:02.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"Ship the recent-index flow"}]}}"#
+                    .to_owned(),
+                r#"{"timestamp":"2026-03-20T00:00:03.000Z","type":"response_item","payload":{"type":"message","role":"assistant","content":[{"type":"output_text","text":"Working on the lightweight summary."}]}}"#
+                    .to_owned(),
+            ],
+        );
+        insert_thread_row(
+            &state_database,
+            "session-001",
+            &session_file,
+            "desktop",
+            &workspace_path,
+            1_742_428_803,
+        );
+
+        let items = load_recent_session_index_from_disk().expect("recent index should load");
+        let item = items
+            .into_iter()
+            .next()
+            .expect("recent index entry should exist");
+
+        assert_eq!(item.session_id, "session-001");
+        assert_eq!(item.display_name, "demo-app");
+        assert_eq!(item.model.as_deref(), Some("gpt-5"));
+        assert_eq!(item.title, "Ship the recent-index flow");
+        assert_eq!(item.status, "done");
+        assert_eq!(
+            item.last_event_summary,
+            "Working on the lightweight summary."
+        );
+        assert_eq!(
+            item.first_user_message.as_deref(),
+            Some("Ship the recent-index flow")
+        );
+    }
+
+    #[test]
+    fn skips_conductor_recent_index_entries() {
+        let ctx = RecentSessionTestContext::new("recent-index-conductor");
+        let workspace_path = ctx
+            .temp_root
+            .join("conductor/workspaces/React-Dashboard/kyiv");
+        let session_file = ctx.sessions_root.join("conductor.jsonl");
+        let state_database = ctx.codex_home.join("state_2.sqlite");
+
+        fs::create_dir_all(&workspace_path).expect("workspace path should exist");
+        create_state_database(&state_database, &[]);
+        write_session_lines(
+            &session_file,
+            vec![
+                session_meta_line("session-conductor", &workspace_path),
+                r#"{"timestamp":"2026-03-20T00:00:01.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"Should never appear"}]}}"#
+                    .to_owned(),
+            ],
+        );
+        insert_thread_row(
+            &state_database,
+            "session-conductor",
+            &session_file,
+            "desktop",
+            &workspace_path,
+            1_742_428_801,
+        );
+
+        let items = load_recent_session_index_from_disk().expect("recent index should load");
+
+        assert!(items.is_empty());
+    }
+
+    #[test]
+    fn skips_recent_index_entries_when_origin_workspace_is_missing() {
+        let ctx = RecentSessionTestContext::new("recent-index-missing-origin");
+        let workspace_path = ctx.temp_root.join("tmp/ghost-workspace");
+        let session_file = ctx.sessions_root.join("missing-origin.jsonl");
+        let state_database = ctx.codex_home.join("state_3.sqlite");
+
+        fs::create_dir_all(&workspace_path).expect("workspace path should exist");
+        create_state_database(&state_database, &[]);
+        write_session_lines(
+            &session_file,
+            vec![
+                session_meta_line("session-missing-origin", &workspace_path),
+                r#"{"timestamp":"2026-03-20T00:00:01.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"This workspace should be rejected"}]}}"#
+                    .to_owned(),
+            ],
+        );
+        insert_thread_row(
+            &state_database,
+            "session-missing-origin",
+            &session_file,
+            "desktop",
+            &workspace_path,
+            1_742_428_801,
+        );
+
+        let items = load_recent_session_index_from_disk().expect("recent index should load");
+
+        assert!(items.is_empty());
+    }
+
+    #[test]
+    fn keeps_recent_automation_sessions_when_workspace_identity_is_valid() {
+        let ctx = RecentSessionTestContext::new("recent-index-automation");
+        let workspace_path = ctx.projects_root.join("automation-demo");
+        let session_file = ctx.sessions_root.join("automation.jsonl");
+        let state_database = ctx.codex_home.join("state_4.sqlite");
+
+        create_git_workspace(&workspace_path);
+        create_state_database(&state_database, &[]);
+        write_session_lines(
+            &session_file,
+            vec![
+                session_meta_line("automation-session", &workspace_path),
+                r#"{"timestamp":"2026-03-20T00:00:01.000Z","type":"turn_context","payload":{"model":"gpt-5"}}"#
+                    .to_owned(),
+                r#"{"timestamp":"2026-03-20T00:00:02.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"Automation: Daily Diary Automation\n# Daily Diary\nAI 에이전트 활동 로그와 Obsidian 볼트 변경사항을 종합한다."}]}}"#
+                    .to_owned(),
+                r#"{"timestamp":"2026-03-20T00:00:03.000Z","type":"event_msg","payload":{"type":"task_complete","last_agent_message":"done"}}"#
+                    .to_owned(),
+            ],
+        );
+        insert_thread_row(
+            &state_database,
+            "automation-session",
+            &session_file,
+            "desktop",
+            &workspace_path,
+            1_742_428_803,
+        );
+
+        let items = load_recent_session_index_from_disk().expect("recent index should load");
+        let item = items
+            .into_iter()
+            .next()
+            .expect("automation session should remain visible");
+
+        assert_eq!(item.session_id, "automation-session");
+        assert_eq!(item.display_name, "automation-demo");
+        assert_eq!(item.title, DEFAULT_THREAD_TITLE);
+        assert!(item.first_user_message.is_none());
+    }
+
+    #[test]
+    fn skips_exec_recent_index_entries() {
+        let ctx = RecentSessionTestContext::new("recent-index-exec");
+        let workspace_path = ctx.projects_root.join("demo-app");
+        let session_file = ctx.sessions_root.join("exec-session.jsonl");
+        let state_database = ctx.codex_home.join("state_5.sqlite");
+
+        create_git_workspace(&workspace_path);
+        create_state_database(&state_database, &[]);
+        write_session_lines(
+            &session_file,
+            vec![
+                session_meta_line_with_source("session-exec", &workspace_path, "exec"),
+                r#"{"timestamp":"2026-03-20T00:00:02.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"This exec session should stay hidden"}]}}"#
+                    .to_owned(),
+            ],
+        );
+        insert_thread_row(
+            &state_database,
+            "session-exec",
+            &session_file,
+            "exec",
+            &workspace_path,
+            1_742_428_802,
+        );
+
+        let items = load_recent_session_index_from_disk().expect("recent index should load");
+
+        assert!(items.is_empty());
+    }
+
+    #[test]
+    fn keeps_recent_index_entries_from_existing_codex_worktrees() {
+        let ctx = RecentSessionTestContext::new("recent-index-worktree-live");
+        let origin_workspace = ctx.projects_root.join("Obsidian-frontend-journey");
+        let workspace_path = ctx
+            .temp_root
+            .join(".codex/worktrees/5594/Obsidian-frontend-journey");
+        let session_file = ctx.sessions_root.join("live-worktree-session.jsonl");
+        let state_database = ctx.codex_home.join("state_6.sqlite");
+
+        create_linked_worktree(&origin_workspace, &workspace_path, "5594");
+        create_state_database(&state_database, &[]);
+        write_session_lines(
+            &session_file,
+            vec![
+                session_meta_line_with_source("session-worktree", &workspace_path, "vscode"),
+                r#"{"timestamp":"2026-03-20T00:00:02.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"Automation: Dialy Diary Automation\n# Daily Diary\nCodex app live sessions should keep valid worktrees visible."}]}}"#
+                    .to_owned(),
+                r#"{"timestamp":"2026-03-20T00:00:03.000Z","type":"response_item","payload":{"type":"message","role":"assistant","content":[{"type":"output_text","text":"Prepared the daily diary draft."}]}}"#
+                    .to_owned(),
+            ],
+        );
+        insert_thread_row(
+            &state_database,
+            "session-worktree",
+            &session_file,
+            "vscode",
+            &workspace_path,
+            1_742_428_802,
+        );
+
+        let item = load_recent_session_index_from_disk()
+            .expect("recent index should load")
+            .into_iter()
+            .next()
+            .expect("existing worktree session should remain visible");
+
+        assert_eq!(item.session_id, "session-worktree");
+        assert_eq!(item.display_name, "Obsidian-frontend-journey");
+        assert_eq!(item.origin_path, origin_workspace.display().to_string());
+    }
+
+    #[test]
+    fn skips_stale_codex_worktree_recent_index_entries() {
+        let ctx = RecentSessionTestContext::new("recent-index-worktree-stale");
+        let workspace_path = ctx.temp_root.join(".codex/worktrees/6971/exem-ui");
+        let session_file = ctx.sessions_root.join("stale-worktree-session.jsonl");
+        let state_database = ctx.codex_home.join("state_6.sqlite");
+
+        create_state_database(&state_database, &[]);
+        write_session_lines(
+            &session_file,
+            vec![
+                session_meta_line_with_source("session-stale-worktree", &workspace_path, "vscode"),
+                r#"{"timestamp":"2026-03-20T00:00:02.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"This stale worktree session should stay hidden"}]}}"#
+                    .to_owned(),
+            ],
+        );
+        insert_thread_row(
+            &state_database,
+            "session-stale-worktree",
+            &session_file,
+            "vscode",
+            &workspace_path,
+            1_742_428_802,
+        );
+
+        let items = load_recent_session_index_from_disk().expect("recent index should load");
+
+        assert!(items.is_empty());
+    }
+
+    #[test]
+    fn collapses_nested_project_workspace_and_skips_environment_context_titles() {
+        let ctx = RecentSessionTestContext::new("recent-index-nested-project");
+        let workspace_root = ctx.projects_root.join("exem-ui");
+        let nested_workspace = workspace_root.join("packages/ui");
+        let session_file = ctx.sessions_root.join("nested-project.jsonl");
+        let state_database = ctx.codex_home.join("state_7.sqlite");
+
+        create_git_workspace(&workspace_root);
+        fs::create_dir_all(&nested_workspace).expect("nested workspace path should exist");
+        create_state_database(&state_database, &[]);
+        write_session_lines(
+            &session_file,
+            vec![
+                session_meta_line("session-nested", &nested_workspace),
+                r##"{"timestamp":"2026-03-20T00:00:01.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"# AGENTS.md instructions for /Users/choegihwan/Documents/Projects/exem-ui"}]}}"##
+                    .to_owned(),
+                r##"{"timestamp":"2026-03-20T00:00:02.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"<environment_context> <cwd>/Users/choegihwan/Documents/Projects/exem-ui/packages/ui</cwd> <shell>zsh</shell> </environment_context>"}]}}"##
+                    .to_owned(),
+                r##"{"timestamp":"2026-03-20T00:00:03.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"List the instruction sources you loaded."}]}}"##
+                    .to_owned(),
+            ],
+        );
+        insert_thread_row(
+            &state_database,
+            "session-nested",
+            &session_file,
+            "desktop",
+            &nested_workspace,
+            1_742_428_803,
+        );
+
+        let item = load_recent_session_index_from_disk()
+            .expect("recent index should load")
+            .into_iter()
+            .next()
+            .expect("recent index entry should exist");
+
+        assert_eq!(item.display_name, "exem-ui");
+        assert_eq!(item.origin_path, workspace_root.display().to_string());
+        assert_eq!(item.title, "List the instruction sources you loaded.");
+        assert_eq!(
+            item.first_user_message.as_deref(),
+            Some("List the instruction sources you loaded.")
+        );
+    }
+
+    #[test]
+    fn skips_recent_index_entries_from_archived_projects_workspace_paths() {
+        let ctx = RecentSessionTestContext::new("recent-index-archived-projects");
+        let workspace_path = ctx
+            .temp_root
+            .join("Documents/Archives/Projects/mfo_v5_starter");
+        let session_file = ctx.sessions_root.join("archived-project.jsonl");
+        let state_database = ctx.codex_home.join("state_8.sqlite");
+
+        create_git_workspace(&workspace_path);
+        create_state_database(&state_database, &[]);
+        write_session_lines(
+            &session_file,
+            vec![
+                session_meta_line_with_source("session-archived-project", &workspace_path, "vscode"),
+                r#"{"timestamp":"2026-03-20T00:00:02.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"Archived workspace should stay hidden"}]}}"#
+                    .to_owned(),
+            ],
+        );
+        insert_thread_row(
+            &state_database,
+            "session-archived-project",
+            &session_file,
+            "vscode",
+            &workspace_path,
+            1_742_428_802,
+        );
+
+        let items = load_recent_session_index_from_disk().expect("recent index should load");
+
+        assert!(items.is_empty());
+    }
+
+    #[test]
+    fn skips_boot_only_recent_threads_without_meaningful_activity() {
+        let ctx = RecentSessionTestContext::new("recent-index-boot-only");
+        let workspace_path = ctx.projects_root.join("exem-ui");
+        let session_file = ctx.sessions_root.join("boot-only.jsonl");
+        let state_database = ctx.codex_home.join("state_10.sqlite");
+
+        create_git_workspace(&workspace_path);
+        create_state_database(&state_database, &[]);
+        write_session_lines(
+            &session_file,
+            vec![session_meta_line_with_source(
+                "session-boot-only",
+                &workspace_path,
+                "cli",
+            )],
+        );
+        insert_thread_row(
+            &state_database,
+            "session-boot-only",
+            &session_file,
+            "cli",
+            &workspace_path,
+            1_742_428_800,
+        );
+
+        let items = load_recent_session_index_from_disk().expect("recent index should load");
+
+        assert!(items.is_empty());
+    }
+
+    #[test]
+    fn excludes_archived_threads_from_recent_index() {
+        let ctx = RecentSessionTestContext::new("recent-index-archived");
+        let workspace_path = ctx.projects_root.join("demo-app");
+        let archived_session_file = ctx.sessions_root.join("archived-live.jsonl");
+        let visible_session_file = ctx.sessions_root.join("visible-live.jsonl");
+        let state_database = ctx.codex_home.join("state_9.sqlite");
+
+        create_git_workspace(&workspace_path);
+        create_state_database(&state_database, &[]);
+        write_session_lines(
+            &archived_session_file,
+            vec![
+                session_meta_line("session-archived", &workspace_path),
+                r#"{"timestamp":"2026-03-20T00:00:01.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"Archived thread should stay hidden"}]}}"#
+                    .to_owned(),
+            ],
+        );
+        write_session_lines(
+            &visible_session_file,
+            vec![
+                session_meta_line("session-visible", &workspace_path),
+                r#"{"timestamp":"2026-03-20T00:00:01.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"Visible thread should remain"}]}}"#
+                    .to_owned(),
+            ],
+        );
+        insert_thread_row_with_archive_flag(
+            &state_database,
+            "session-archived",
+            &archived_session_file,
+            "desktop",
+            &workspace_path,
+            1_742_428_801,
+            true,
+        );
+        insert_thread_row(
+            &state_database,
+            "session-visible",
+            &visible_session_file,
+            "desktop",
+            &workspace_path,
+            1_742_428_802,
+        );
+
+        let items = load_recent_session_index_from_disk().expect("recent index should load");
+
+        assert_eq!(items.len(), 1);
+        assert_eq!(items[0].session_id, "session-visible");
+    }
+
+    #[test]
+    fn rejects_archived_threads_from_recent_snapshot_loading() {
+        let ctx = RecentSessionTestContext::new("recent-snapshot-archived");
+        let workspace_path = ctx.projects_root.join("demo-app");
+        let session_file = ctx.sessions_root.join("archived-selected.jsonl");
+        let state_database = ctx.codex_home.join("state_11.sqlite");
+
+        create_git_workspace(&workspace_path);
+        create_state_database(&state_database, &[]);
+        write_session_lines(
+            &session_file,
+            vec![
+                session_meta_line("session-archived", &workspace_path),
+                r#"{"timestamp":"2026-03-20T00:00:01.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"Archived thread should not hydrate"}]}}"#
+                    .to_owned(),
+            ],
+        );
+        insert_thread_row_with_archive_flag(
+            &state_database,
+            "session-archived",
+            &session_file,
+            "desktop",
+            &workspace_path,
+            1_742_428_801,
+            true,
+        );
+
+        let snapshot =
+            load_recent_session_snapshot_from_disk(session_file.to_string_lossy().as_ref());
+
+        assert!(snapshot.is_none());
+    }
+
+    #[test]
+    fn loads_selected_recent_snapshot_and_matching_subagents_only() {
+        let ctx = RecentSessionTestContext::new("recent-detail");
+        let workspace_path = ctx.projects_root.join("demo-app");
+        let selected_file = ctx.sessions_root.join("selected.jsonl");
+        let matched_subagent_file = ctx.sessions_root.join("matched-sub.jsonl");
+        let unrelated_subagent_file = ctx.sessions_root.join("other-sub.jsonl");
+        let state_database = ctx.codex_home.join("state_12.sqlite");
+
+        create_git_workspace(&workspace_path);
+        create_state_database(&state_database, &[]);
+        write_session_lines(
+            &selected_file,
+            vec![
+                session_meta_line_with_fork("session-001", &workspace_path, Some("fork-001")),
+                r#"{"timestamp":"2026-03-20T00:00:01.000Z","type":"turn_context","payload":{"model":"gpt-5"}}"#
+                    .to_owned(),
+                r#"{"timestamp":"2026-03-20T00:00:02.000Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"Inspect only this session"}]}}"#
+                    .to_owned(),
+            ],
+        );
+        write_session_lines(
+            &matched_subagent_file,
+            vec![
+                r#"{"timestamp":"2026-03-20T00:01:00.000Z","type":"session_meta","payload":{"id":"sub-001","source":{"subagent":{"thread_spawn":{"parent_thread_id":"session-001","depth":1,"agent_nickname":"Euler","agent_role":"worker"}}},"cwd":"/tmp/test","timestamp":"2026-03-20T00:01:00.000Z"}}"#.to_owned(),
+                r#"{"timestamp":"2026-03-20T00:01:01.000Z","type":"turn_context","payload":{"model":"gpt-5-mini"}}"#.to_owned(),
+                r#"{"timestamp":"2026-03-20T00:01:02.000Z","type":"response_item","payload":{"type":"message","role":"assistant","content":[{"type":"output_text","text":"Matched child"}]}}"#.to_owned(),
+            ],
+        );
+        write_session_lines(
+            &unrelated_subagent_file,
+            vec![
+                r#"{"timestamp":"2026-03-20T00:02:00.000Z","type":"session_meta","payload":{"id":"sub-999","source":{"subagent":{"thread_spawn":{"parent_thread_id":"other-parent","depth":1,"agent_nickname":"Noether","agent_role":"worker"}}},"cwd":"/tmp/test","timestamp":"2026-03-20T00:02:00.000Z"}}"#.to_owned(),
+                r#"{"timestamp":"2026-03-20T00:02:02.000Z","type":"response_item","payload":{"type":"message","role":"assistant","content":[{"type":"output_text","text":"Unrelated child"}]}}"#.to_owned(),
+            ],
+        );
+        insert_thread_row(
+            &state_database,
+            "session-001",
+            &selected_file,
+            "desktop",
+            &workspace_path,
+            1_742_428_802,
+        );
+
+        let snapshot =
+            load_recent_session_snapshot_from_disk(selected_file.to_string_lossy().as_ref())
+                .expect("selected recent snapshot should exist");
+
+        assert_eq!(snapshot.session_id, "session-001");
+        assert_eq!(snapshot.subagents.len(), 1);
+        assert_eq!(snapshot.subagents[0].session_id, "sub-001");
+        assert_eq!(snapshot.subagents[0].agent_nickname, "Euler");
+    }
+}
