+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod sessions;
+pub(crate) mod workspace;
