+++ src-tauri/src/domain/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod ingest_policy;
+pub(crate) mod session;
+pub(crate) mod workspace;
