+++ src-tauri/src/infrastructure/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod filesystem;
+pub(crate) mod git;
+pub(crate) mod session_jsonl;
+pub(crate) mod state_sqlite;
