+++ src-tauri/src/state/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod archive_cache;
