+++ src-tauri/src/support/error.rs
@@ -1,0 +1,5 @@
+use std::io;
+
+pub(crate) fn map_sqlite_error(error: rusqlite::Error) -> io::Error {
+    io::Error::other(error)
+}
