+++ src-tauri/src/support/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod error;
+pub(crate) mod text;
